fix: separate validations to different files and update tests
