+++ tests/mocks/ingestionValidator.ts
@@ -1,0 +1,24 @@
+import { IngestionValidator } from '../../src/layers/models/ingestionValidator';
+
+const validateGpkgFilesMock = jest.fn();
+const validateGdalInfoMock = jest.fn();
+const validateSourceDirectoryMock = jest.fn();
+const validateNotWatchDirMock = jest.fn();
+const fileValidatorValidateExistsMock = jest.fn();
+
+const ingestionValidatorMock = {
+  validateGpkgFiles: validateGpkgFilesMock,
+  validateGdalInfo: validateGdalInfoMock,
+  validateSourceDirectory: validateSourceDirectoryMock,
+  validateNotWatchDir: validateNotWatchDirMock,
+  validateExists: fileValidatorValidateExistsMock,
+} as unknown as IngestionValidator;
+
+export {
+  validateGpkgFilesMock,
+  validateGdalInfoMock,
+  validateSourceDirectoryMock,
+  validateNotWatchDirMock,
+  fileValidatorValidateExistsMock,
+  ingestionValidatorMock,
+};
